expanded actuator manager structure

+++ include/OVTopics/common_types.hpp
@@ -6,7 +6,7 @@ namespace Actuators {
 
 /* Generic Motion Trajectory */
 typedef struct {
-    uint8_t* traj_arr;
+    uint32_t* traj_arr;
     uint32_t traj_arr_length;
     uint32_t sample_time;
     bool cyclical; /* If cyclical: must be a closed path -> end = start */
